requirements/base.txt
click==4.0
Django==1.8.2
jsonfield==1.0.3
git+https://github.com/powersurge360/django-multi-gtfs@feature/django-1.7-upgrade
psycopg2==2.6.1
wheel==0.24.0
python-dotenv==0.1.2
factory-boy==2.5.2
djangorestframework-gis==0.9
django-filter==0.10.0
freezegun==0.3.4
